Fix: negative energy demanded from heat source
When the temperature required by the radiator is much lower than the
previous timestep and the energy demand is lower than the energy
required to warm the emitter the emitter module demands negative energy
from the heat source. To fix this the energy required from the current
timestep has been set to 0.