change storey input
- Current there is a single "storey" input which was the number of
storeys in a house or the storey of a flat within the building. The
notional dwelling spec requires the number of storeys in the building in
both cases, so the inputs were changed so that the number of storeys in
the building is always asked for and the storey of the dwelling within
the building is only applicable for flats.
- update project file
- update demo files