Fix: energy demand is based on the tank layers the heat source can heat
Previously, although the switching on and off of the heat source
accounted for the positions of the heater and thermostat within the hot
water tank, the energy requirement was being calculated based on the
entire tank. As the heat source can't heat any layers below the heater
position, this demand can never be satisfied so the heat source would
constantly see this demand, leading to the calculated energy consumption
being too high.